Refactor: use search hook within separate Link component
